+++ CHANGELOG.md
@@ -17,6 +17,12 @@ See [Semver](http://semver.org/).
 - 🐞 Backwards-compatible bug fixes
 - 📦 Minor packaging changes
 
+## v5.8.0
+
+- 🐞 Fixed issue where immediately closing a process before a pending task
+  completed resulted in `Error: onExit(exit) called end()`
+- 📦 Updated deps
+
 ## v5.7.1
 
 - 📦 `BatchCluster.end()` should return a `Deferred<void>`